requirements.txt
playwright
google-genai
python-dotenv
pandas
supabase
streamlit
streamlit_option_menu